ne_rabotaet

+++ src/main/java/ru/kata/spring/boot_security/demo/service/UserServiceImpl.java
@@ -25,6 +25,7 @@ public class UserServiceImpl implements UserService {
         return userRepository.findAll();
     }
 
+
     @Override
     public User getUserById(Long id) {
         return userRepository.findById(id).orElse(null);
@@ -48,6 +49,10 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void deleteUser(Long id) {
-        userRepository.deleteById(id);
+        if (userRepository.existsById(id)) {
+            userRepository.deleteById(id);
+        } else {
+            throw new RuntimeException("User not found with id: " + id);
+        }
     }
 }
